pagosApi/requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
mercadopago==2.2.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
